describe commit in about 50 characters and imperative
skip a line, then describe the commit in lines of at most 80 characters.
describe the logic behind the commit and reasoning for choosing a particular
implementation.
in practice most commits are sufficiently described by the first line.

i filled the file with lorem ipsum text because it's a standard filler.
this also gives some context for diffs in subsequent commits.